refactor: 카드 정보 각 필드별 컴포넌트화
